Improve jumping sword attacks
Comments:
-Keep "moving" hero when using jumping sword attack. The hero is not
"free" (not possible while sword attack for engine limitations), so a
built-in movement is used for this.
-Allow to use sword several times on air during the jump. This was not
trivial: since a movement is applied, the engine (for some unknown
reason) does not allow to use the sword during another unfinished sword
attack, so the movement is not directly attached to the hero as a
workaround.